+++ application/controllers/JenisKertas.php
@@ -8,55 +8,71 @@ class JenisKertas extends CI_Controller {
 	}
 
 	public function lst() {
-		$data['page_title'] = 'List Jenis Kertas';
-		$data['page_note'] = 'Daftar seluruh barang dan jenis-jenis kertasnya';
+		if ($_SESSION['printer']['user']['role'] == 0){
+			$data['page_title'] = 'List Jenis Kertas';
+			$data['page_note'] = 'Daftar seluruh barang dan jenis-jenis kertasnya';
 
-		$data['row'] = $this->MJenisKertas->getAllWithParent();
+			$data['row'] = $this->MJenisKertas->getAllWithParent();
 
-		$this->load->view('header', $data);
-		$this->load->view('jeniskertas_list', $data);
-		$this->load->view('footer', $data);
+			$this->load->view('header', $data);
+			$this->load->view('jeniskertas_list', $data);
+			$this->load->view('footer', $data);
+		} else {
+			redirect('login');
+		}
 	}
 
 	public function add() {
-		$data['page_title'] = 'Tambah Jenis Kertas Baru';
-		$data['edited'] = false;
+		if ($_SESSION['printer']['user']['role'] == 0){
+			$data['page_title'] = 'Tambah Jenis Kertas Baru';
+			$data['edited'] = false;
 
-		$data['barang'] = $this->MBarang->getAll();
+			$data['barang'] = $this->MBarang->getAll();
 
-		if ($this->input->post('button') == 'save'){
-			$result = $this->MJenisKertas->insert($this->input->post('barang'),$this->input->post('nama'));
-			if ($result)
-				redirect('jeniskertas');
+			if ($this->input->post('button') == 'save'){
+				$result = $this->MJenisKertas->insert($this->input->post('barang'),$this->input->post('nama'));
+				if ($result)
+					redirect('jeniskertas');
+			}
+
+			$this->load->view('header', $data);
+			$this->load->view('jeniskertas_edit', $data);
+			$this->load->view('footer', $data);
+		} else {
+			redirect('login');
 		}
-
-		$this->load->view('header', $data);
-		$this->load->view('jeniskertas_edit', $data);
-		$this->load->view('footer', $data);
 	}
 
 	public function edit($id) {
-		$data['page_title'] = 'Edit Jenis Kertas';
-		$data['edited'] = true;
+		if ($_SESSION['printer']['user']['role'] == 0){
+			$data['page_title'] = 'Edit Jenis Kertas';
+			$data['edited'] = true;
 
-		$data['barang'] = $this->MBarang->getAll();
+			$data['barang'] = $this->MBarang->getAll();
 
-		$detail = $this->MJenisKertas->getOne($id);
-		$data['detail'] = $detail[0];
+			$detail = $this->MJenisKertas->getOne($id);
+			$data['detail'] = $detail[0];
 
-		if ($this->input->post('button') == 'save'){
-			$result = $this->MJenisKertas->update($id,$this->input->post('barang'),$this->input->post('nama'));
-			if ($result)
-				redirect('jeniskertas');
+			if ($this->input->post('button') == 'save'){
+				$result = $this->MJenisKertas->update($id,$this->input->post('barang'),$this->input->post('nama'));
+				if ($result)
+					redirect('jeniskertas');
+			}
+
+			$this->load->view('header', $data);
+			$this->load->view('jeniskertas_edit', $data);
+			$this->load->view('footer', $data);
+		} else {
+			redirect('login');
 		}
-
-		$this->load->view('header', $data);
-		$this->load->view('jeniskertas_edit', $data);
-		$this->load->view('footer', $data);
 	}
 
 	public function delete($id) {
-		$result = $this->MJenisKertas->delete($id);
-		if ($result) redirect('jeniskertas');
+		if ($_SESSION['printer']['user']['role'] == 0){
+			$result = $this->MJenisKertas->delete($id);
+			if ($result) redirect('jeniskertas');
+		} else {
+			redirect('login');
+		}
 	}
 }
